integ test fixes
