+++ Makefile
@@ -23,7 +23,7 @@ DOCKER:=$(shell command -v podman 2> /dev/null)
 endif
 
 
-.PHONY: build image lint clean test integ-test imagepush
+.PHONY: build image lint clean test  imagepush install-numaflow
 
 build: clean
 	CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -v -o ./dist/aws-sqs-source-go-amd64 main.go
@@ -40,7 +40,7 @@ lint:
 
 test:
 	@echo "Running integration tests..."
-	@go test ./pkg/sqs/sqs_test.go
+	@go test ./pkg/sqs/*
 
 imagepush: build
 	docker buildx build --no-cache -t "$(DOCKERIO_ORG)/numaflow-go/aws-sqs-source-go:$(IMAGE_TAG)" --platform $(PLATFORMS) --target $(TARGET) . --push
@@ -66,6 +66,8 @@ test-e2e:
 	$(MAKE) e2eapi-image
 	kubectl -n numaflow-system delete po -lapp.kubernetes.io/component=controller-manager,app.kubernetes.io/part-of=numaflow
 	kubectl -n numaflow-system delete po e2e-api-pod  --ignore-not-found=true
+	kubectl create ns numaflow-system
+	kubectl apply -n numaflow-system -f https://raw.githubusercontent.com/numaproj/numaflow/stable/config/install.yaml
 	cat pkg/e2e/manifests/e2e-api-pod.yaml |  sed 's@quay.io/numaproj/@$(IMAGE_NAMESPACE)/@' | sed 's/:latest/:$(VERSION)/' | kubectl -n numaflow-system apply -f -
 	kubectl apply -f pkg/e2e/manifests/moto.yaml
 	export AWS_ACCESS_KEY="your_access_key"
@@ -73,7 +75,7 @@ test-e2e:
 	export AWS_SECRET="your_secret"
 	export AWS_QUEUE="your_queue_name"
 	export AWS_ENDPOINT="http://127.0.0.1:5000"
-	go test -v -timeout 15m -count 1 --tags test -p 1 ./pkg/e2e/sqs_e2e_test.go
+	go test -v -timeout 15m -count 1 --tags test -p 1 ./pkg/e2e/test/sqs_e2e_test.go
 	$(MAKE) cleanup-e2e
 
 .PHONY: e2eapi-image
@@ -82,4 +84,5 @@ e2eapi-image: clean dist/e2eapi
 
 clean:
 	-rm -rf ${CURRENT_DIR}/dist
+
