+++ pkg/e2e/test/testdata/sqs_source.yaml
@@ -4,7 +4,7 @@ metadata:
   name: aws-sqs-source
 spec:
   limits:
-    readBatchSize: 10 # for aws sqs the batch size should be between 1 -  10
+    readBatchSize: 10 # for aws sqs the batch size should be between 1 -  10 - SQS Valid values: 1 to 10. Default: 1
   vertices:
     - name: in
       scale:
